rename class
https://coveord.atlassian.net/browse/CDX-361
