+++ packages/cli/src/lib/snapshot/reportViewer/reportViewer.ts
@@ -5,7 +5,7 @@ import {
 } from '@coveord/platform-client';
 import {cli} from 'cli-ux';
 import {red, italic} from 'chalk';
-import {ReportViewerSection} from './reportViewerResourceSection';
+import {ReportViewerSection} from './reportViewerSection';
 import {ReportViewerStyles} from './reportViewerStyles';
 import {
   ReportViewerOperationName,
